refactor(settings): share ProjectSearchInput across teams and projects tabs
